First complete pass at adding all of the ViewsService methods.
These methods provide for the ability to retrieve and manipulate metadata about datasets and views, as well as to query for datasets and views that match given search criteria. It also allows for the creation, read, update, and deletion of rows of data from datasets and views.

Second change relates to how to handle various data formats. Instead of trying to guess what the developer might want to do with data from some format, the idea instead is: Everything will be called for is via JSON by default an we'll decode into an object or array. Otherwise, if they want XML or XLS or whatever, they're on their own.